Add some subcommand and some templates

+++ lib/goa_model_gen/generator.rb
@@ -7,15 +7,24 @@ require "active_support/core_ext/string"
 
 module GoaModelGen
   class Generator
-    def initialize(template_path)
-      content = File
-      @erb = ERB.new(File.read(template_path), nil, "-")
-      @erb.filename = template_path
+    # These are used in templates
+    attr_reader :config
+    attr_accessor :types
+
+    def initialize(config)
+      @config = config
     end
 
-    def run(types, path)
-      content = @erb.result(binding)
+    def run(rel_path, path, overwrite: false)
+      return if File.exist?(path) && !overwrite
+      abs_path = File.expand_path('../' + rel_path, __FILE__)
+      erb = ERB.new(File.read(abs_path), nil, "-")
+      erb.filename = abs_path
+      content = erb.result(binding)
       open(path, 'w'){|f| f.puts(content) }
+      if (File.extname(path) == '.go') && !config.gofmt_disabled
+        system("gofmt -w #{path}")
+      end
     end
 
   end
